MB-67564: Fix narrowing conversion in Histogram
When we compile Magma with MSVC with `/we4267 /we4244` enabled, we see
the following error:

```
C:\...\platform\include\platform/histogram.h(157):
error C4244: '=': conversion from 'T' to 'double',
possible loss of data
        with
        [
            T=size_t
        ]
```

This patch aims to fix this by using a `gsl::narrow_cast`.

Reviewed-on: https://review.couchbase.org/c/platform/+/231522
Tested-by: Build Bot <build@couchbase.com>

+++ include/platform/histogram.h
@@ -19,6 +19,7 @@
 #include <memory>
 #include <numeric>
 #include <vector>
+#include <gsl/gsl-lite.hpp>
 
 // Custom microseconds duration used for measuring histogram stats.
 // Stats will never be negative, so an unsigned Rep is used.
@@ -154,7 +155,7 @@ public:
         : _growth(growth),
           _start(start) {
         // Dividing two durations returns a value of the underlying Rep.
-        _width = width / T(1);
+        _width = gsl::narrow_cast<double>(width / T(1));
     }
 
     /**
